+++ openeo/rest/datacube.py
@@ -2403,10 +2403,16 @@ class DataCube(_ProcessGraphAbstraction):
             of another :py:class:`~openeo.rest.datacube.DataCube` instance.
         """
         if self._connection:
-            formats = set(self._connection.list_output_formats().keys())
+            output_formats = self._connection.list_output_formats()
+            formats = set(output_formats)
             # TODO: map format to correct casing too?
             if format.lower() not in {f.lower() for f in formats}:
                 raise ValueError("Invalid format {f!r}. Should be one of {s}".format(f=format, s=formats))
+            self._connection._warn_on_invalid_output_format_options(
+                format=format,
+                options=options,
+                output_formats=output_formats,
+            )
 
         pg = self._build_pgnode(
             process_id="save_result",