+++ openeo/rest/vectorcube.py
@@ -211,11 +211,15 @@ class VectorCube(_ProcessGraphAbstraction):
             returns a :py:class:`~openeo.rest.result.SaveResult` instance instead
             of another :py:class:`~openeo.rest.vectorcube.VectorCube` instance.
         """
+        format = format or "GeoJSON"
+        if self._connection:
+            self._connection._warn_on_invalid_output_format_options(format=format, options=options)
+
         pg = self._build_pgnode(
             process_id="save_result",
             arguments={
                 "data": self,
-                "format": format or "GeoJSON",
+                "format": format,
                 # TODO: leave out options if unset?
                 "options": options or {},
             },